fixed minor compiler warnings in probe.cu

+++ src/gpu/VirtualFluids_GPU/Visitor/Probe.cu
@@ -200,7 +200,7 @@ void Probe::init(Parameter* para, GridProvider* gridProvider, CudaMemoryManager*
             probeParams[level]->quantitiesH[var] = true;
             probeParams[level]->arrayOffsetsH[var] = arrOffset;
 
-            arrOffset += getPostProcessingVariableNames(static_cast<PostProcessingVariable>(var)).size();
+            arrOffset += uint(getPostProcessingVariableNames(static_cast<PostProcessingVariable>(var)).size());
 
         }}
         
@@ -400,7 +400,7 @@ void Probe::writeGridFiles(Parameter* para, int level, std::vector<std::string>&
         {
             PostProcessingVariable quantity = static_cast<PostProcessingVariable>(var);
             real coeff;
-            uint n_arrs = getPostProcessingVariableNames(quantity).size();
+            uint n_arrs = uint(getPostProcessingVariableNames(quantity).size());
 
             switch(quantity)
             {
